get_data/scraper/requirements.txt
black
beautifulsoup4
requests
pytest
pandas
google-cloud-storage
flask
gunicorn